tests: Move cli tests to test_settings.py
